pyproject.toml
[tool.poetry]
name = "dbtai"
version = "0.1.0"
description = "`dbtai` is a utility CLI command to generate dbt model documentation for a given model using OpenAI."
authors = ["Henning Holgersen"]
keywords = [
    "documentation",
    "dbt",
    "openai"
]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = "<3.14,>=3.8.0"
openai = ">1.1.0"
click = "^8.1.3"
"ruamel.yaml" = "^0.18.6"
inquirer = "^3.2.4"
appdirs = "^1.4.4"
pyyaml = "^6.0.1"

[tool.poetry.dev-dependencies]
pytest = "^8.0.1"
tox = "^3.24.4"
flake8 = "^3.9.2"
black = "^22.9"
pydocstyle = "^6.1.1"
mypy = "^0.910"
types-requests = "^2.26.1"
isort = "^5.10.1"

[tool.isort]
profile = "black"
multi_line_output = 3 # Vertical Hanging Indent
src_paths = "dbtai"

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# CLI declaration
dbtai = 'dbtai.cli:dbtai'